Move worker files loader out of services
