+++ lib/sidekiq_adhoc_job/worker_files_loader.rb
